pyproject.toml
[tool.poetry]
name = "cmssmm"
version = "0.1.0"
description = "All in one content management and social media management for Streamers and Content Creators"
authors = ["Rich Kevan <rskevan@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
django-ninja = "^1.1.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
pillow = "^10.3.0"
boto3 = "^1.34.97"
django-storages = "^1.14.2"
django-tailwind = "^3.8.0"
django-browser-reload = "^1.12.1"
hurry-filesize = "^0.9"
exifread = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
